LayoutTests/platform/chromium/http/tests/security/listener/xss-inactive-closure-expected.txt

CONSOLE MESSAGE: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
This tests that when a frame navigates to a new page, closures in the old page cannot access page content of the new page if there are from different domains.

You should see PASSED at the end of page if the test passes.

FAILED: document is undefined
  
